Moved AuthQueryHandler from memgraph.cpp to glue subdirectory
