+++ src/glue/CMakeLists.txt
@@ -1,0 +1,4 @@
+set(mg_glue_sources auth.cpp auth_handler.cpp communication.cpp)
+
+add_library(mg-glue STATIC ${mg_glue_sources})
+target_link_libraries(mg-glue mg-query mg-auth)
